Investigate planner theta* slowdown
